openWorldQuest "Perfect World Government"
Instead of one map we create a open world with quests aka role playing game. It is a bit like Half Life and a bit like Fallout and little like Left for Dead. 

Story is you grow up in a peaceful childhood. Democracy seems to be the ultimate freedom and peace. But simultaneously  one war after the next one is declared. But you all believe it. 
Everyone beliefs it. 

However you discover unanswered questions ? Who is behind all this. The world is changing - but only for a view people. The mass majority won't ever get it.